flightbooking/requirements.txt
Django==4.1.3
psycopg2==2.9.2
django-jazzmin==2.6.0
Django-Gtts==0.4
